+++ spec/kettle/dev/reset_cli_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe Kettle::Dev::ResetCLI do
   it "checks Gemfile.lock without changing it" do
     File.write(File.join(@root, "Gemfile.lock"), <<~LOCK)
       GEM
-        remote: https://gem.coop/
+        remote: https://rubygems.org/
         specs:
           rake (13.4.2)
 
@@ -33,6 +33,7 @@ RSpec.describe Kettle::Dev::ResetCLI do
     path = File.join(@root, "Gemfile.lock")
     allow(Kettle::Dev::LockfileReset).to receive(:new).and_return(resetter)
     allow(resetter).to receive(:lockfile_paths_for).with("Gemfile.lock").and_return([path])
+    allow(resetter).to receive(:release_lockfiles_target?).with("Gemfile.lock").and_return(false)
     allow(resetter).to receive(:normalization_needed?).with(path).and_return(true)
     allow(resetter).to receive(:reset).with("Gemfile.lock").and_return(path)
 
@@ -45,9 +46,22 @@ RSpec.describe Kettle::Dev::ResetCLI do
     path = File.join(@root, "Gemfile.lock")
     allow(Kettle::Dev::LockfileReset).to receive(:new).and_return(resetter)
     allow(resetter).to receive(:lockfile_paths_for).with("Gemfile.lock").and_return([path])
+    allow(resetter).to receive(:release_lockfiles_target?).with("Gemfile.lock").and_return(false)
     allow(resetter).to receive(:normalization_needed?).with(path).and_return(false)
 
     expect(resetter).not_to receive(:reset)
+    expect(cli.run!).to eq(0)
+  end
+
+  it "always rewrites release lockfiles" do
+    cli = described_class.new(%w[release-lockfiles], root: @root)
+    resetter = instance_double(Kettle::Dev::LockfileReset)
+    path = File.join(@root, "Gemfile.lock")
+    allow(Kettle::Dev::LockfileReset).to receive(:new).and_return(resetter)
+    allow(resetter).to receive(:lockfile_paths_for).with("release-lockfiles").and_return([path])
+    allow(resetter).to receive(:release_lockfiles_target?).with("release-lockfiles").and_return(true)
+    allow(resetter).to receive(:reset).with("release-lockfiles").and_return([path])
+
     expect(cli.run!).to eq(0)
   end
 end